Trying mv to go/src directly.
