+++ cmd/cmd_suite_test.go
@@ -1,8 +1,9 @@
-package cmd_test
+package cmd
 
 import (
 	"testing"
 
+	"github.com/8Mobius8/go-habits/api"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -11,3 +12,19 @@ func TestCmd(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Cmd Suite")
 }
+
+var _ = Describe("Status", func() {
+	It("returns the funnies when Habitica is reachable.", func() {
+		var resp api.Status
+		resp.Status = "up"
+
+		Expect(StatusMessage(resp)).To(Equal("Habitica is reachable, GO catch all those pets!"))
+	})
+
+	It("returns the sad when Habitica is unreachable.", func() {
+		var resp api.Status
+		resp.Status = "down"
+
+		Expect(StatusMessage(resp)).To(Equal(":( Habitica is unreachable."))
+	})
+})
