--- cmd/status_test.go
@@ -1,24 +1,0 @@
-package cmd_test
-
-import (
-	"github.com/8Mobius8/go-habits/api"
-	. "github.com/8Mobius8/go-habits/cmd"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-var _ = Describe("Status", func() {
-	It("returns the funnies when Habitica is reachable.", func() {
-		var resp api.Status
-		resp.Status = "up"
-
-		Expect(StatusMessage(resp)).To(Equal("Habitica is reachable, GO catch all those pets!"))
-	})
-
-	It("returns the sad when Habitica is unreachable.", func() {
-		var resp api.Status
-		resp.Status = "down"
-
-		Expect(StatusMessage(resp)).To(Equal(":( Habitica is unreachable."))
-	})
-})
